+++ modules/html-storybook/src/BgBox.stories.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import BgBox from '@reykjavik/hanna-react/BgBox';
+
+import { StoryComponent, StoryParameters } from './storytypes';
+
+export default {
+  title: 'BgBox',
+  parameters: {
+    knobs: { disabled: false },
+  } as StoryParameters,
+};
+
+export const _BgBox: StoryComponent = () => (
+  <BgBox>
+    <p>Some content that needs a "background box".</p>
+  </BgBox>
+);